pyproject.toml
[project]
name = "FlexFL"
version = "0.1.0"
description = "Resilient Federated Learning Framework"
readme = "README.md"
authors = [
    { name = "Leonardo Almeida", email = "leonardoalmeida7777@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "kaggle>=1.7.4.2",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "permetrics>=2.0.0",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "wget>=3.2",
]


[project.scripts]
flexfl = "flexfl.cli.fl:main"
flexfl-division = "flexfl.cli.division:main"
flexfl-preprocess = "flexfl.cli.preprocess:main"
flexfl-res = "flexfl.cli.auto_restart:main"


[project.optional-dependencies]
all = [
    "tensorflow[and-cuda]>=2.19.0",
    "kafka-python>=2.1.2",
    "mpi4py>=4.0.3",
    "paho-mqtt>=2.1.0",
    "eclipse-zenoh==1.1.1",
]
tf = [
    "tensorflow[and-cuda]>=2.19.0",
]
kafka = [
    "kafka-python>=2.1.2",
]
mpi = [
    "mpi4py>=4.0.3",
]
mqtt = [
    "paho-mqtt>=2.1.0",
]
zenoh = [
    "eclipse-zenoh==1.1.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
